backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
ipwhois==1.3.0
python-dotenv==1.0.1
asyncio-cache==0.0.1
aiofiles==24.1.0
requests==2.32.3
netaddr==1.3.0
